updated file structure to be more organized, updated home page, added gallery and archive component skeletons
