+++ src/components/Gallery.jsx
@@ -1,0 +1,12 @@
+// Gallery.jsx
+import React from 'react';
+
+export const Gallery = () => {
+  // Add logic to fetch and display gallery images
+  return (
+    <div>
+      <h2>Gallery</h2>
+      {/* Render gallery images */}
+    </div>
+  );
+};
